requirements.txt
openai
python-dotenv
mysql-connector-python
